docs/requirements.txt
# docs
myst-nb==0.10.1
markdown-it-py==0.6.2
MarkupSafe==1.1.1
Sphinx==3.5.3
sphinx-autoapi==1.7.0
sphinx-copybutton==0.3.1
sphinx-material==0.0.32
sphinx-panels==0.5.2
sphinx-togglebutton==0.2.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-dotnetdomain==0.4
sphinxcontrib-golangdomain==0.2.0.dev0
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-napoleon==0.7
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
six==1.15.0
